@@ -32,23 +32,23 @@
   },
   "peerDependencies": {
     "graphql": "^15.5.0 || ^16.0.0",
-    "react": "^16.8.0 || ^17.0.0 || ^18.0.0",
-    "react-dom": "^16.8.0 || ^17.0.0 || ^18.0.0"
+    "react": "^16.8.0 || ^17 || ^18",
+    "react-dom": "^16.8.0 || ^17 || ^18"
   },
   "dependencies": {
+    "@headlessui/react": "^1.7.14",
+    "@radix-ui/react-dialog": "^1.0.3",
+    "@radix-ui/react-visually-hidden": "^1.0.2",
+    "@radix-ui/react-tooltip": "^1.0.5",
+    "@radix-ui/react-dropdown-menu": "^2.0.4",
     "@graphiql/toolkit": "^0.8.4",
-    "@reach/combobox": "^0.17.0",
-    "@reach/dialog": "^0.17.0",
-    "@reach/listbox": "^0.17.0",
-    "@reach/menu-button": "^0.17.0",
-    "@reach/tooltip": "^0.17.0",
-    "@reach/visually-hidden": "^0.17.0",
     "clsx": "^1.2.1",
     "codemirror": "^5.65.3",
     "codemirror-graphql": "^2.0.8",
     "copy-to-clipboard": "^3.2.0",
     "graphql-language-service": "^5.1.6",
     "markdown-it": "^12.2.0",
+    "framer-motion": "^6.5.1",
     "set-value": "^4.1.0"
   },
   "devDependencies": {
@@ -59,8 +59,8 @@
     "@vitejs/plugin-react": "^1.3.0",
     "graphql": "^16.4.0",
     "postcss-nesting": "^10.1.7",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "typescript": "^4.6.3",
     "vite": "^2.9.13",
     "vite-plugin-react-svg": "^0.2.0"
