@@ -57,8 +57,8 @@
   },
   "peerDependencies": {
     "graphql": "^15.5.0 || ^16.0.0",
-    "react": "^16.8.0 || ^17.0.0 || ^18.0.0",
-    "react-dom": "^16.8.0 || ^17.0.0 || ^18.0.0"
+    "react": "^16.8.0 || ^17 || ^18",
+    "react-dom": "^16.8.0 || ^17 || ^18"
   },
   "devDependencies": {
     "@cypress/webpack-preprocessor": "^5.5.0",
@@ -87,8 +87,8 @@
     "postcss-import": "15.1.0",
     "postcss-preset-env": "^8.0.1",
     "prop-types": "15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hot-loader": "^4.12.20",
     "react-test-renderer": "^16.13.1",
     "require-context.macro": "^1.2.2",
