@@ -1,5 +1,6 @@
-import { forwardRef } from 'react';
+import { forwardRef, ReactNode } from 'react';
 import { clsx } from 'clsx';
+import { Reorder } from 'framer-motion';
 import { CloseIcon } from '../icons';
 import { createComponentGroup } from '../utility/component-group';
 import { UnStyledButton } from './button';
@@ -9,26 +10,29 @@ import './tabs.css';
 
 type TabProps = {
   isActive?: boolean;
+  value: object;
+  className?: string;
+  children: ReactNode;
 };
 
-const TabRoot = forwardRef<
-  HTMLDivElement,
-  TabProps & JSX.IntrinsicElements['div']
->(({ isActive, ...props }, ref) => (
-  <div
-    {...props}
-    ref={ref}
-    role="tab"
-    aria-selected={isActive}
-    className={clsx(
-      'graphiql-tab',
-      isActive && 'graphiql-tab-active',
-      props.className,
-    )}
-  >
-    {props.children}
-  </div>
-));
+const TabRoot = forwardRef<HTMLLIElement, TabProps>(
+  ({ isActive, value, children, className, ...props }, ref) => (
+    <Reorder.Item
+      {...props}
+      ref={ref}
+      value={value}
+      aria-selected={isActive ? 'true' : undefined}
+      role="tab"
+      className={clsx(
+        'graphiql-tab',
+        isActive && 'graphiql-tab-active',
+        className,
+      )}
+    >
+      {children}
+    </Reorder.Item>
+  ),
+);
 TabRoot.displayName = 'Tab';
 
 const TabButton = forwardRef<
@@ -68,16 +72,26 @@ export const Tab = createComponentGroup(TabRoot, {
   Close: TabClose,
 });
 
-export const Tabs = forwardRef<HTMLDivElement, JSX.IntrinsicElements['div']>(
-  (props, ref) => (
-    <div
+type TabsProps = {
+  values: object[];
+  onReorder: (newOrder: any[]) => void;
+  className?: string;
+  children: ReactNode;
+};
+
+export const Tabs = forwardRef<HTMLUListElement, TabsProps>(
+  ({ values, onReorder, children, className, ...props }, ref) => (
+    <Reorder.Group
       {...props}
       ref={ref}
+      values={values}
+      onReorder={onReorder}
+      axis="x"
       role="tablist"
-      className={clsx('graphiql-tabs', props.className)}
+      className={clsx('graphiql-tabs', className)}
     >
-      {props.children}
-    </div>
+      {children}
+    </Reorder.Group>
   ),
 );
 Tabs.displayName = 'Tabs';
